+++ aldrin/src/reply.rs
@@ -11,13 +11,13 @@ use std::task::{Context, Poll};
 /// Future to await the result of a call.
 #[must_use = "futures do nothing unless you `.await` or poll them"]
 pub struct Reply<T, E> {
-    recv: Receiver<CallFunctionResult>,
+    recv: Receiver<Result<CallFunctionResult, Error>>,
     function: u32,
     phantom: PhantomData<fn() -> (T, E)>,
 }
 
 impl<T, E> Reply<T, E> {
-    pub(crate) fn _new(recv: Receiver<CallFunctionResult>, function: u32) -> Self {
+    pub(crate) fn new(recv: Receiver<Result<CallFunctionResult, Error>>, function: u32) -> Self {
         Self {
             recv,
             function,
@@ -30,28 +30,37 @@ impl<T: Deserialize, E: Deserialize> Future for Reply<T, E> {
     type Output = Result<Result<T, E>, Error>;
 
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output> {
-        let res = match Pin::new(&mut self.recv).poll(cx) {
-            Poll::Ready(Ok(res)) => res,
-            Poll::Ready(Err(_)) => return Poll::Ready(Err(Error::Shutdown)),
-            Poll::Pending => return Poll::Pending,
-        };
-
-        Poll::Ready(match res {
-            CallFunctionResult::Ok(t) => match t.deserialize() {
-                Ok(t) => Ok(Ok(t)),
-                Err(e) => Err(Error::invalid_reply(e)),
+        match Pin::new(&mut self.recv).poll(cx) {
+            Poll::Ready(Ok(Ok(CallFunctionResult::Ok(t)))) => match t.deserialize() {
+                Ok(t) => Poll::Ready(Ok(Ok(t))),
+                Err(e) => Poll::Ready(Err(Error::invalid_reply(e))),
             },
 
-            CallFunctionResult::Err(e) => match e.deserialize() {
-                Ok(e) => Ok(Err(e)),
-                Err(e) => Err(Error::invalid_reply(e)),
+            Poll::Ready(Ok(Ok(CallFunctionResult::Err(e)))) => match e.deserialize() {
+                Ok(e) => Poll::Ready(Ok(Err(e))),
+                Err(e) => Poll::Ready(Err(Error::invalid_reply(e))),
             },
 
-            CallFunctionResult::Aborted => Err(Error::CallAborted),
-            CallFunctionResult::InvalidService => Err(Error::InvalidService),
-            CallFunctionResult::InvalidFunction => Err(Error::invalid_function(self.function)),
-            CallFunctionResult::InvalidArgs => Err(Error::invalid_arguments(self.function, None)),
-        })
+            Poll::Ready(Ok(Ok(CallFunctionResult::Aborted))) => {
+                Poll::Ready(Err(Error::CallAborted))
+            }
+
+            Poll::Ready(Ok(Ok(CallFunctionResult::InvalidService))) => {
+                Poll::Ready(Err(Error::InvalidService))
+            }
+
+            Poll::Ready(Ok(Ok(CallFunctionResult::InvalidFunction))) => {
+                Poll::Ready(Err(Error::invalid_function(self.function)))
+            }
+
+            Poll::Ready(Ok(Ok(CallFunctionResult::InvalidArgs))) => {
+                Poll::Ready(Err(Error::invalid_arguments(self.function, None)))
+            }
+
+            Poll::Ready(Ok(Err(e))) => Poll::Ready(Err(e)),
+            Poll::Ready(Err(_)) => Poll::Ready(Err(Error::Shutdown)),
+            Poll::Pending => Poll::Pending,
+        }
     }
 }