+++ aldrin/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `low_level::Event::new()`.
 - Add `EventListener::is_finished()`.
 - Add `low_level::Proxy`, which acts as a new basis for generated proxy types. It handles function
+  calls and events.
+- Add `Reply<T, E>`, which replaces both `PendingFunctionResult<T, E>` and
+  `PendingFunctionValue<T>`.
 
 ### Changed
 
@@ -19,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename `Events` to `EventListener` and move it to the `low_level` module.
 - Rename `BusListener::finished()` to `is_finished()`.
 - Rename `Discoverer::finished()` to `is_finished()`.
+- `Reply<T, E>` replaces both `PendingFunctionResult<T, E>` and `PendingFunctionValue<T>`.
+- Remove `Handle::call_function()`. Calls can now only be made through `low_level::Proxy::call()`.
 
 ## [0.4.0] - 2024-03-21